moving non-templated functions in cpp
